Change compiler warnings to be both GCC and Clang compatible
Removed unused variable warnings, by removing them or adding [[maybe_unused]]

Replace variable-length arrays with std::vector

Make librevcpu.cc #include "SST.h" with warnings suppressed in SST headers

Remove unnecessary casts of std::stoull and std::stoi

Simplify unnecessarily complicated vector offset calculation

-      uint32_t Off = 0;
-      if( Addr == baseAddr[i] ){
-        Off = 0;  // lets avoid division by zero
-      }else{
-        Off = (uint32_t)((Addr-baseAddr[i])/4);
-      }
+      uint32_t Off = static_cast<uint32_t>((Addr-baseAddr[i])/4);

(There is no division BY zero; it's division OF zero by 4.)

+++ src/RevOpts.cc
@@ -86,7 +86,7 @@ bool RevOpts::InitStartAddrs( std::vector<std::string> StartAddrs ){
     if( vstr[0] == "CORES" ){
       // set all cores to the target machine model
       std::string::size_type sz = 0;
-      uint64_t Addr = (uint64_t)(std::stoull(vstr[1], &sz, 0));
+      uint64_t Addr = std::stoull(vstr[1], &sz, 0);
       for( unsigned i=0; i<numCores; i++ ){
         startAddr.find(i)->second = Addr;
       }
@@ -100,12 +100,12 @@ bool RevOpts::InitStartAddrs( std::vector<std::string> StartAddrs ){
     if( vstr.size() != 2 )
       return false;
 
-    unsigned Core = (unsigned)(std::stoi(vstr[0], nullptr, 0));
+    unsigned Core = std::stoi(vstr[0], nullptr, 0);
     if( Core > numCores )
       return false;
 
     std::string::size_type sz = 0;
-    uint64_t Addr = (uint64_t)(std::stoull(vstr[1], &sz, 0));
+    uint64_t Addr = std::stoull(vstr[1], &sz, 0);
 
     startAddr.find(Core)->second = Addr;
     vstr.clear();
@@ -121,7 +121,7 @@ bool RevOpts::InitStartSymbols( std::vector<std::string> StartSymbols ){
     if( vstr.size() != 2 )
       return false;
 
-    unsigned Core = (unsigned)(std::stoi(vstr[0], nullptr, 0));
+    unsigned Core = std::stoi(vstr[0], nullptr, 0);
     if( Core > numCores )
       return false;
 
@@ -157,7 +157,7 @@ bool RevOpts::InitMachineModels( std::vector<std::string> Machines ){
     if( vstr.size() != 2 )
       return false;
 
-    unsigned Core = (unsigned)(std::stoi(vstr[0], nullptr, 0));
+    unsigned Core = std::stoi(vstr[0], nullptr, 0);
     if( Core > numCores )
       return false;
 
@@ -175,7 +175,7 @@ bool RevOpts::InitInstTables( std::vector<std::string> InstTables ){
     if( vstr.size() != 2 )
       return false;
 
-    unsigned Core = (unsigned)(std::stoi(vstr[0], nullptr, 0));
+    unsigned Core = std::stoi(vstr[0], nullptr, 0);
     if( Core > numCores )
       return false;
 
@@ -194,9 +194,9 @@ bool RevOpts::InitMemCosts( std::vector<std::string> MemCosts ){
     if( vstr.size() != 3 )
       return false;
 
-    unsigned Core = (unsigned)(std::stoi(vstr[0], nullptr, 0));
-    unsigned Min  = (unsigned)(std::stoi(vstr[1], nullptr, 0));
-    unsigned Max  = (unsigned)(std::stoi(vstr[2], nullptr, 0));
+    unsigned Core = std::stoi(vstr[0], nullptr, 0);
+    unsigned Min  = std::stoi(vstr[1], nullptr, 0);
+    unsigned Max  = std::stoi(vstr[2], nullptr, 0);
     memCosts[Core].first  = Min;
     memCosts[Core].second = Max;
     if( (Min==0) || (Max==0) ){
